Update so that unauthenticated client can access public buckets (#24)
This pull request adds logic that handles three different authentication methods for Google Cloud Storage. It provides a fallback mechanism to unauthenticated mode if no default credentials are found.

If a JSON key path is provided, it uses it to initialize normally. If not, it tries to use the default application credentials, and if they don't exist, it falls back to unauthenticated mode. The code also sets the header provider and retry settings for the storage.

Co-authored-by: Ugo Sangiorgi <ugo.sangiorgi@gmail.com>